+++ src/components/FormModal.tsx
@@ -4,11 +4,11 @@ import autoBindMethods from 'class-autobind-decorator';
 
 import * as Antd from 'antd';
 
-import FormManager from './FormManager';
-import { ICardConfig } from './interfaces';
+import FormManager from '../utilities/FormManager';
+import { ICardConfig } from '../interfaces';
 
-import FormFields from './FormFields';
-import { fillInFieldSets, getFieldSetFields } from './common';
+import FormFields from '../building-blocks/FormFields';
+import { fillInFieldSets, getFieldSetFields } from '../utilities/common';
 import { computed } from 'mobx';
 
 interface IProps {
