+++ src/index.ts
@@ -1,24 +1,25 @@
 // Components
-export { default as ArrayCard } from './ArrayCard';
-export { default as EditableArrayCard } from './EditableArrayCard';
-export { default as Card } from './Card';
-export { default as Cards } from './Cards';
-export { default as EditableCard } from './EditableCard';
-export { default as FormModal } from './FormModal';
-export { default as SummaryCard } from './SummaryCard';
+export { default as ArrayCard } from './components/ArrayCard';
+export { default as Card } from './components/Card';
+export { default as Cards } from './components/Cards';
+export { default as EditableArrayCard } from './components/EditableArrayCard';
+export { default as EditableCard } from './components/EditableCard';
+export { default as FormCard } from './components/SummaryCard';
+export { default as FormModal } from './components/FormModal';
+export { default as SummaryCard } from './components/SummaryCard';
 
 // Utility classes and functions
-export { default as FormManager } from './FormManager';
+export { default as FormManager } from './utilities/FormManager';
 export {
   fillInFieldConfig,
   fillInFieldSets,
   getFieldSetFields,
   isFieldSetSimple,
-} from './common';
+} from './utilities/common';
 
 export * from './interfaces';
 
 // Lower-level building blocks and helper components
-export { default as FormFields } from './FormFields';
-export { default as OptionSelectDisplay, formatOptionSelect } from './OptionSelectDisplay';
-export { default as Info, Label, Value, CARD_COL_LABEL, CARD_COL_VALUE } from './Info';
+export { default as FormFields } from './building-blocks/FormFields';
+export { default as OptionSelectDisplay, formatOptionSelect } from './inputs/OptionSelectDisplay';
+export { default as Info, Label, Value, CARD_COL_LABEL, CARD_COL_VALUE } from './building-blocks/Info';
